Update coloring for my Leet code solutions Style
New coloring.sed file is added, this is generated by awk file. It is used to instead of _sedtmp.

coloring.sh now could handle multiple files. It also handled path issue

+++ src/coloring.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+if [[ -z $1 ]]; then
+    echo "No files specified, please pass file as parameter." >&2
+    exit
+fi
 
 #resource files
 sedfile="coloring.sed"
@@ -7,14 +11,31 @@ regexfile="cpp.regex"
 htmlencode="htmlencode.sed"
 
 
-
-tmpfile[0]=".encoded$" #encoded temp file
-tmpfile[1]=".body$"
+tmpfile[0]=".encoded$"  #encoded temp file
+tmpfile[1]=".body$"     #preprocessed body
 #tmpfile[2]=".html$"
 
-file="sample.cpp"
+hastmp=0
+directory=${1%/*}
+
+if [[ $directory = $1 ]]; then
+    directory=""
+else
+    directory=$directory"/"
+fi
+
+files=$(ls $1)
+
+for file in $files; do
+    
+    file=${file##*/} #remove path info
+
+    if [[ -f $directory$file ]]; then
+
+#file="sample.cpp"
 filename=${file%.*} #remove file extension
-htmlfile=$filename".html"
+file=$directory$file #concatenate file with directory
+htmlfile=$directory$filename".html"
 
 header="<html>
     <head>
@@ -42,7 +63,7 @@ bottom='</code>
 </html>'
 
 #htmlencode
-sed -n -E -f $htmlencode sample.cpp | cat > ${tmpfile[0]}
+sed -n -E -f $htmlencode $file | cat > ${tmpfile[0]}
 
 #create sed substitution actions
 awk -f coloring.awk -v s=$sedfile -v r=$regexfile ${tmpfile[0]} > ${tmpfile[1]};
@@ -54,5 +75,14 @@ sed -n -E -f $sedfile ${tmpfile[1]} >> $htmlfile
 #append bottom
 echo $bottom >> $htmlfile
 
+echo "Done coloring for: "$file;
+hastmp=1
+    fi
+    
+done
+
 #remove all temp files
-rm ${tmpfile[@]}
+if [[ $hastmp > 0 ]]; then
+    rm ${tmpfile[@]}
+fi
+
